fix/Исправил замечания ревью
