+++ src/components/what-we-do-authors/index.ts
