+++ src/components/what-we-do-desc/what-we-do-desc.module.css
@@ -108,7 +108,7 @@
   }
 
   .imgSelected {
-    order: 2;
+    order: -1;
   }
 
   .containerSelected {
